+++ sieve/sieve_bespoke.c
@@ -6,9 +6,9 @@
 
 #include "parameters.h"
 
-size_t sieve(uint64_t *primes, size_t len) {
+size_t sieve(uint32_t *primes, size_t len) {
   size_t p = 0;
-  uint64_t i = 2;
+  uint32_t i = 2;
   while (p < len) {
     bool divisible = false;
     for (size_t j = 0; j < p; j++) {
@@ -26,8 +26,8 @@ size_t sieve(uint64_t *primes, size_t len) {
 }
 
 int main(void) {
-  const size_t primes_len = sizeof(reference) / sizeof(uint64_t);
-  uint64_t primes[primes_len];
+  const size_t primes_len = sizeof(reference) / sizeof(uint32_t);
+  uint32_t primes[primes_len];
   size_t num_primes = sieve(primes, primes_len);
 
   print_primes(primes, num_primes);
